+++ README.md
@@ -1,10 +1,10 @@
-# PSR, defining a native persistence container protocol
+# Remote method invokation library implementation
 
-[![Latest Stable Version](https://poser.pugx.org/appserver-io-psr/persistencecontainerprotocol/v/stable.png)](https://packagist.org/packages/appserver-io-psr/persistencecontainerprotocol) [![Total Downloads](https://poser.pugx.org/appserver-io-psr/persistencecontainerprotocol/downloads.png)](https://packagist.org/packages/appserver-io-psr/persistencecontainerprotocol) [![Latest Unstable Version](https://poser.pugx.org/appserver-io-psr/persistencecontainerprotocol/v/unstable.png)](https://packagist.org/packages/appserver-io-psr/persistencecontainerprotocol) [![License](https://poser.pugx.org/appserver-io-psr/persistencecontainerprotocol/license.png)](https://packagist.org/packages/appserver-io-psr/persistencecontainerprotocol) [![Build Status](https://travis-ci.org/appserver-io-psr/persistencecontainerprotocol.png)](https://travis-ci.org/appserver-io-psr/persistencecontainerprotocol)[![Scrutinizer Code Quality](https://scrutinizer-ci.com/g/appserver-io-psr/persistencecontainerprotocol/badges/quality-score.png?b=master)](https://scrutinizer-ci.com/g/appserver-io-psr/persistencecontainerprotocol/?branch=master)[![Code Coverage](https://scrutinizer-ci.com/g/appserver-io-psr/persistencecontainerprotocol/badges/coverage.png?b=master)](https://scrutinizer-ci.com/g/appserver-io-psr/persistencecontainerprotocol/?branch=master)
+[![Latest Stable Version](https://poser.pugx.org/appserver-io/rmi/v/stable.png)](https://packagist.org/packages/appserver-io/rmi) [![Total Downloads](https://poser.pugx.org/appserver-io/rmi/downloads.png)](https://packagist.org/packages/appserver-io/rmi) [![Latest Unstable Version](https://poser.pugx.org/appserver-io/rmi/v/unstable.png)](https://packagist.org/packages/appserver-io/rmi) [![License](https://poser.pugx.org/appserver-io/rmi/license.png)](https://packagist.org/packages/appserver-io/rmi) [![Build Status](https://travis-ci.org/appserver-io/rmi.png)](https://travis-ci.org/appserver-io/rmi)[![Scrutinizer Code Quality](https://scrutinizer-ci.com/g/appserver-io/rmi/badges/quality-score.png?b=master)](https://scrutinizer-ci.com/g/appserver-io/rmi/?branch=master)[![Code Coverage](https://scrutinizer-ci.com/g/appserver-io/rmi/badges/coverage.png?b=master)](https://scrutinizer-ci.com/g/appserver-io/rmi/?branch=master)
 
 ## Introduction
 
-Protocol to enable a [server](<https://github.com/appserver-io/server>) to act as part of a persistence service.
+Library that implements basic remote method invocation.
 
 ## Installation
 
@@ -13,7 +13,7 @@ If you want to use the library with your application you can install it by addin
 ```sh
 {
     "require": {
-        "appserver-io-psr/persistencecontainerprotocol": "dev-master"
+        "appserver-io/rmi": "dev-master"
     },
 }
 ```
