+++ bootstrap.php
@@ -11,15 +11,14 @@
  *
  * PHP version 5
  *
- * @category   Appserver
- * @package    Psr
- * @subpackage PersistenceContainerProtocol
- * @author     Tim Wagner <tw@appserver.io>
- * @copyright  2014 TechDivision GmbH <info@appserver.io>
- * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link       https://github.com/appserver-io-psr/persistencecontainerprotocol
- * @link       http://www.appserver.io
+ * @category  Library
+ * @package   Psr
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2014 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/rmi
+ * @link      http://www.appserver.io
  */
 
 $loader = require 'vendor/autoload.php';
-$loader->add('AppserverIo\\Psr\\PersistenceContainerProtocol', 'src');
+$loader->add('AppserverIo\\RemoteMethodInvocation', 'src');
