+++ build.default.properties
@@ -8,4 +8,4 @@
 #--------------------------------------------------------------------------------
 
 # ---- Module Release Settings --------------------------------------------------
-release.version = 0.2.9
+release.version = 0.3.0
