+++ build.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <!DOCTYPE project>
-<project name="appserver-io-psr/persistencecontainerprotocol" default="composer-init" basedir=".">
+<project name="appserver-io/rmi" default="composer-init" basedir=".">
     
     <!-- initialize ENV variable -->
     <property environment="env" />
