+++ phpunit.xml
@@ -1,6 +1,6 @@
 <phpunit>
     <testsuites>
-	    <testsuite name="appserver-io-psr/persistencecontainerprotocol PHPUnit testsuite">
+	    <testsuite name="appserver-io/rmi PHPUnit testsuite">
 	        <directory suffix="Test.php">tests</directory>
 	    </testsuite>
 	</testsuites>
