--- src/AppserverIo/Psr/PersistenceContainerProtocol/BeanContext.php
@@ -1,133 +1,0 @@
-<?php
-
-/**
- * AppserverIo\Psr\PersistenceContainerProtocol\BeanContext
- *
- * NOTICE OF LICENSE
- *
- * This source file is subject to the Open Software License (OSL 3.0)
- * that is available through the world-wide-web at this URL:
- * http://opensource.org/licenses/osl-3.0.php
- *
- * PHP version 5
- *
- * @category   Appserver
- * @package    Psr
- * @subpackage PersistenceContainerProtocol
- * @author     Tim Wagner <tw@appserver.io>
- * @copyright  2014 TechDivision GmbH <info@appserver.io>
- * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link       https://github.com/appserver-io-psr/persistencecontainerprotocol
- * @link       http://www.appserver.io
- */
-
-namespace AppserverIo\Psr\PersistenceContainerProtocol;
-
-/**
- * The interface for the bean manager registered for the application.
- *
- * @category   Appserver
- * @package    Psr
- * @subpackage PersistenceContainerProtocol
- * @author     Tim Wagner <tw@appserver.io>
- * @copyright  2014 TechDivision GmbH <info@appserver.io>
- * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link       https://github.com/appserver-io-psr/persistencecontainerprotocol
- * @link       http://www.appserver.io
- */
-interface BeanContext
-{
-
-    /**
-     * The unique identifier to be registered in the application context.
-     *
-     * @var string
-     */
-    const IDENTIFIER = 'BeanContext';
-
-    /**
-     * Returns the absolute path to the web application.
-     *
-     * @return string The absolute path
-     */
-    public function getWebappPath();
-
-    /**
-     * Tries to locate the queue that handles the request and returns the instance
-     * if one can be found.
-     *
-     * @param \AppserverIo\Psr\PersistenceContainerProtocol\RemoteMethod $remoteMethod The remote method call
-     * @param array                                                      $args         The arguments passed to the session beans constructor
-     *
-     * @return object The requested bean instance
-     */
-    public function locate(RemoteMethod $remoteMethod, array $args = array());
-
-    /**
-     * Runs a lookup for the session bean with the passed class name and
-     * session ID.
-     *
-     * If the passed class name is a session bean an instance
-     * will be returned.
-     *
-     * @param string $className The name of the session bean's class
-     * @param string $sessionId The session ID
-     * @param array  $args      The arguments passed to the session beans constructor
-     *
-     * @return object The requested bean instance
-     * @throws \TechDivision\PersistenceContainer\InvalidBeanTypeException Is thrown if passed class name is no session bean or is a entity bean (not implmented yet)
-     */
-    public function lookup($className, $sessionId = null, array $args = array());
-
-    /**
-     * Attaches the passed bean, depending on it's type to the container.
-     *
-     * @param object $instance  The bean instance to attach
-     * @param string $sessionId The session-ID when we have stateful session bean
-     *
-     * @return void
-     * @throws \Exception Is thrown if we have a stateful session bean, but no session-ID passed
-     */
-    public function attach($instance, $sessionId = null);
-
-    /**
-     * Returns a new reflection class intance for the passed class name.
-     *
-     * @param string $className The class name to return the reflection class instance for
-     *
-     * @return \AppserverIo\Lang\Reflection\ReflectionClass The reflection instance
-     */
-    public function newReflectionClass($className);
-
-    /**
-     * Returns a reflection class intance for the passed class name.
-     *
-     * @param string $className The class name to return the reflection class instance for
-     *
-     * @return \AppserverIo\Lang\Reflection\ReflectionClass The reflection instance
-     * @see \DependencyInjectionContainer\Interfaces\ProviderInterface::getReflectionClass()
-     */
-    public function getReflectionClass($className);
-
-    /**
-     * Returns a reflection class intance for the passed class name.
-     *
-     * @param object $instance The instance to return the reflection class instance for
-     *
-     * @return \AppserverIo\Lang\Reflection\ReflectionClass The reflection instance
-     * @see \DependencyInjectionContainer\Interfaces\ProviderInterface::newReflectionClass()
-     * @see \DependencyInjectionContainer\Interfaces\ProviderInterface::getReflectionClass()
-     */
-    public function getReflectionClassForObject($instance);
-
-    /**
-     * Returns a new instance of the passed class name.
-     *
-     * @param string      $className The fully qualified class name to return the instance for
-     * @param string|null $sessionId The session-ID, necessary to inject stateful session beans (SFBs)
-     * @param array       $args      Arguments to pass to the constructor of the instance
-     *
-     * @return object The instance itself
-     */
-    public function newInstance($className, $sessionId = null, array $args = array());
-}
