+++ src/AppserverIo/RemoteMethodInvocation/RemoteMethodCallException.php
@@ -1,0 +1,38 @@
+<?php
+
+/**
+ * AppserverIo\RemoteMethodInvocation\RemoteMethodCallException
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @category  Library
+ * @package   Rmi
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2014 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/rmi
+ * @link      http://www.appserver.io
+ */
+
+namespace AppserverIo\RemoteMethodInvocation;
+
+/**
+ * This is the exception that is thrown if a remote method call fails.
+ *
+ * @category  Library
+ * @package   Rmi
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2014 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/rmi
+ * @link      http://www.appserver.io
+ */
+class RemoteMethodCallException extends \Exception
+{
+}
