+++ src/AppserverIo/RemoteMethodInvocation/RemoteObject.php
@@ -1,0 +1,52 @@
+<?php
+
+/**
+ * AppserverIo\RemoteMethodInvocation\RemoteObject
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @category  Library
+ * @package   Rmi
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2014 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/rmi
+ * @link      http://www.appserver.io
+ */
+
+namespace AppserverIo\RemoteMethodInvocation;
+
+/**
+ * Interface for all remote objects.
+ *
+ * @category  Library
+ * @package   Rmi
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2014 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/rmi
+ * @link      http://www.appserver.io
+ */
+interface RemoteObject
+{
+
+    /**
+     * Returns the session instance.
+     *
+     * @return \AppserverIo\RemoteMethodInvocation\Session The session instance
+     */
+    public function getSession();
+
+    /**
+     * The name of the original object.
+     *
+     * @return string The name of the original object
+     */
+    public function getClassName();
+}
