+++ src/AppserverIo/RemoteMethodInvocation/Session.php
@@ -1,0 +1,61 @@
+<?php
+
+/**
+ * AppserverIo\RemoteMethodInvocation\Session
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @category  Library
+ * @package   Rmi
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2014 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/rmi
+ * @link      http://www.appserver.io
+ */
+
+namespace AppserverIo\RemoteMethodInvocation;
+
+/**
+ * The interface for the session.
+ *
+ * @category  Library
+ * @package   Rmi
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2014 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/rmi
+ * @link      http://www.appserver.io
+ */
+interface Session
+{
+
+    /**
+     * Returns the ID of the session to use.
+     *
+     * @return string The session ID
+     */
+    public function getSessionId();
+
+    /**
+     * Invokes the remote method over the connection.
+     *
+     * @param \AppserverIo\RemoteMethodInvocation\RemoteMethod $remoteMethod The remote method call to invoke
+     *
+     * @return mixed the method return value
+     */
+    public function send(RemoteMethod $remoteMethod);
+
+    /**
+     * Creates a remote inital context instance.
+     *
+     * @return \AppserverIo\RemoteMethodInvocation\RemoteObject The proxy for the inital context
+     */
+    public function createInitialContext();
+}
