+++ tests/AppserverIo/RemoteMethodInvocation/RemoteMethodCallTest.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * AppserverIo\Psr\PersistenceContainerProtocol\RemoteMethodCall
+ * AppserverIo\RemoteMethodInvocation\RemoteMethodCall
  *
  * NOTICE OF LICENSE
  *
@@ -11,29 +11,27 @@
  *
  * PHP version 5
  *
- * @category   Appserver
- * @package    Psr
- * @subpackage PersistenceContainerProtocol
- * @author     Tim Wagner <tw@appserver.io>
- * @copyright  2014 TechDivision GmbH <info@appserver.io>
- * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link       https://github.com/appserver-io-psr/persistencecontainerprotocol
- * @link       http://www.appserver.io
+ * @category  Library
+ * @package   Rmi
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2014 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/rmi
+ * @link      http://www.appserver.io
  */
 
-namespace AppserverIo\Psr\PersistenceContainerProtocol;
+namespace AppserverIo\RemoteMethodInvocation;
 
 /**
  * Test implementation for the RemoteMethodCall class.
  *
- * @category   Appserver
- * @package    Psr
- * @subpackage PersistenceContainerProtocol
- * @author     Tim Wagner <tw@appserver.io>
- * @copyright  2014 TechDivision GmbH <info@appserver.io>
- * @license    http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
- * @link       https://github.com/appserver-io-psr/persistencecontainerprotocol
- * @link       http://www.appserver.io
+ * @category  Library
+ * @package   Rmi
+ * @author    Tim Wagner <tw@appserver.io>
+ * @copyright 2014 TechDivision GmbH <info@appserver.io>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/appserver-io/rmi
+ * @link      http://www.appserver.io
  */
 class RemoteMethodCallTest extends \PHPUnit_Framework_TestCase
 {
@@ -48,7 +46,7 @@ class RemoteMethodCallTest extends \PHPUnit_Framework_TestCase
     /**
      * The remote method call instance we want to test.
      *
-     * @var \AppserverIo\Psr\PersistenceContainerProtocol\RemoteMethodCall
+     * @var \AppserverIo\RemoteMethodInvocation\RemoteMethodCall
      */
     protected $remoteMethodCall;
